requirements.txt
python-dotenv
PyPDF2
python-docx
streamlit
langchain-openai
openai
langchain
fastapi